automatically import global vars on document save
Instead of invoking a macro to import the gloval vars once the document
has been saved, make it automatic !
This commit turns the original macro into a mod (i.e. has the same
structure than a workbench), and registers a document observer during
FreeCAD startup. This document observers then calls the global vars
import function when a document is about to be saved, to add the global
vars spreadsheet to it.

So the global vars spreadsheet import is easier for the end user now :
simply create and save a new .FCStd file and when you have a FCProject
file, the global vars are automatically linked to the .FCStd file when
saving it.